requirements.txt
pgmpy>=0.1.25
networkx>=3.0
numpy>=1.24
pandas>=1.5
matplotlib>=3.7
seaborn>=0.12
python-dotenv>=1.0
